% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.pi.permute}
\alias{get.pi.permute}
\title{get the null distribution of the \code{get.pi} function}
\usage{
get.pi.permute(posmat, fun, r = 1, r.low = rep(0, length(r)), permutations)
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{fun}{the function to evaluate}

\item{r}{the series of spatial distances we are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{permutations}{the number of permute iterations}
}
\value{
pi values for all the distances we looked at
}
\description{
Does permutations to calculate the null distribution of get pi
if there were no spatial dependence. Randomly reassigns coordinates
to each observation permutations times
}
\examples{
\dontrun{
 R/examples/get_pi_permute.R
 }
 
}
\seealso{
Other get.pi: 
\code{\link{get.pi.a.clustsurvey.hh.wts}()},
\code{\link{get.pi.a.clustsurvey.wts.bootstrap}()},
\code{\link{get.pi.a.clustsurvey.wts}()},
\code{\link{get.pi.a.typed.clustsurvey.hh.wts}()},
\code{\link{get.pi.a.typed.clustsurvey.wts}()},
\code{\link{get.pi.b.clustsurvey.hh.wts}()},
\code{\link{get.pi.b.clustsurvey.wts.bootstrap}()},
\code{\link{get.pi.b.clustsurvey.wts}()},
\code{\link{get.pi.b.typed.clustsurvey.hh.wts}()},
\code{\link{get.pi.b.typed.clustsurvey.wts}()},
\code{\link{get.pi.bootstrap}()},
\code{\link{get.pi.ci}()},
\code{\link{get.pi.clustsurvey.bootstrap}()},
\code{\link{get.pi.clustsurvey.ci}()},
\code{\link{get.pi.clustsurvey.hh.wts.window}()},
\code{\link{get.pi.clustsurvey.window}()},
\code{\link{get.pi.clustsurvey.wts.window}()},
\code{\link{get.pi.clustsurvey}()},
\code{\link{get.pi.hh.typed}()},
\code{\link{get.pi.typed.bootstrap}()},
\code{\link{get.pi.typed.permute}()},
\code{\link{get.pi.typed.survey}()},
\code{\link{get.pi.typed.wts.bootstrap}()},
\code{\link{get.pi.typed.wts.ci}()},
\code{\link{get.pi.typed.wts.permute}()},
\code{\link{get.pi.typed.wts}()},
\code{\link{get.pi.typed}()},
\code{\link{get.pi}()}
}
\concept{get.pi}
