% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs_clusters.r
\name{get.pi.typed.wts.permute}
\alias{get.pi.typed.wts.permute}
\title{get the null distribution of the \code{get.pi.typed.wts} function}
\usage{
get.pi.typed.wts.permute(
  posmat,
  typeA = -1,
  typeB = -1,
  r = 1,
  r.low = rep(0, length(r)),
  permutations
)
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{typeA}{the "from" type that we are interested in, -1 is wildcard}

\item{typeB}{the "to" type that we are interested i, -1 is wildcard}

\item{r}{the series of spatial distances wer are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{permutations}{the number of permute iterations}
}
\value{
pi values for all the distances we looked at
}
\description{
Does permutations to calculate the null distribution of get pi
if there were no spatial dependence. Randomly reassigns coordinates
to each observation permutations times
}
\seealso{
Other get.pi: 
\code{\link{get.pi.a.clustsurvey.hh.wts}()},
\code{\link{get.pi.a.clustsurvey.wts.bootstrap}()},
\code{\link{get.pi.a.clustsurvey.wts}()},
\code{\link{get.pi.a.typed.clustsurvey.hh.wts}()},
\code{\link{get.pi.a.typed.clustsurvey.wts}()},
\code{\link{get.pi.b.clustsurvey.hh.wts}()},
\code{\link{get.pi.b.clustsurvey.wts.bootstrap}()},
\code{\link{get.pi.b.clustsurvey.wts}()},
\code{\link{get.pi.b.typed.clustsurvey.hh.wts}()},
\code{\link{get.pi.b.typed.clustsurvey.wts}()},
\code{\link{get.pi.bootstrap}()},
\code{\link{get.pi.ci}()},
\code{\link{get.pi.clustsurvey.bootstrap}()},
\code{\link{get.pi.clustsurvey.ci}()},
\code{\link{get.pi.clustsurvey.hh.wts.window}()},
\code{\link{get.pi.clustsurvey.window}()},
\code{\link{get.pi.clustsurvey.wts.window}()},
\code{\link{get.pi.clustsurvey}()},
\code{\link{get.pi.hh.typed}()},
\code{\link{get.pi.permute}()},
\code{\link{get.pi.typed.bootstrap}()},
\code{\link{get.pi.typed.gridcells}()},
\code{\link{get.pi.typed.permute}()},
\code{\link{get.pi.typed.survey}()},
\code{\link{get.pi.typed.wts.bootstrap}()},
\code{\link{get.pi.typed.wts.ci}()},
\code{\link{get.pi.typed.wts}()},
\code{\link{get.pi.typed}()},
\code{\link{get.pi}()}
}
\author{
Justin Lessler and Henrik Salje
}
\concept{get.pi}
