% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs_clusters_window.r
\name{get.tau.clustsurvey.window}
\alias{get.tau.clustsurvey.window}
\title{Optimizewd version of \code{get.tau} for typed data}
\usage{
get.tau.clustsurvey.window(posmat, r = 1, r.low = rep(0, length(r)))
}
\arguments{
\item{posmat}{a matrix with columns type, x, y, and weight}

\item{r}{the series of spatial distances wer are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{typeA}{the "from" type that we are interested in, -1 is wildcard}

\item{typeB}{the "to" type that we are interested i, -1 is wildcard}
}
\value{
tau values for all the distances we looked at
}
\description{
Version of th e \code{get.tau} function that is optimized for
statically typed data. That is dat where we want the relationship between
points of type A and points of type B
}
\seealso{
Other get.tau: 
\code{\link{get.tau.bootstrap}()},
\code{\link{get.tau.ci}()},
\code{\link{get.tau.clustsurvey.bootstrap}()},
\code{\link{get.tau.clustsurvey.ci}()},
\code{\link{get.tau.clustsurvey.hh.wts.window}()},
\code{\link{get.tau.clustsurvey.hh.wts}()},
\code{\link{get.tau.clustsurvey.wts.bootstrap}()},
\code{\link{get.tau.clustsurvey.wts.ci}()},
\code{\link{get.tau.clustsurvey.wts.window}()},
\code{\link{get.tau.clustsurvey.wts}()},
\code{\link{get.tau.clustsurvey}()},
\code{\link{get.tau.hh.typed.bootstrap}()},
\code{\link{get.tau.hh.typed}()},
\code{\link{get.tau.permute}()},
\code{\link{get.tau.typed.bootstrap.parallel}()},
\code{\link{get.tau.typed.bootstrap}()},
\code{\link{get.tau.typed.clustsurvey.hh.wts}()},
\code{\link{get.tau.typed.clustsurvey.wts}()},
\code{\link{get.tau.typed.permute}()},
\code{\link{get.tau.typed.survey}()},
\code{\link{get.tau.typed.wts.bootstrap.parallel}()},
\code{\link{get.tau.typed.wts.bootstrap}()},
\code{\link{get.tau.typed.wts.ci}()},
\code{\link{get.tau.typed.wts.permute}()},
\code{\link{get.tau.typed.wts}()},
\code{\link{get.tau.typed}()},
\code{\link{get.tau}()}
}
\author{
Justin Lessler and Henrik Salje
}
\concept{get.tau}
