% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs_survey.r
\name{get.tau.typed.survey}
\alias{get.tau.typed.survey}
\title{Optimizewd version of \code{get.tau} for typed data}
\usage{
get.tau.typed.survey(
  posmat,
  typeA = -1,
  typeB = -1,
  r = 1,
  r.low = rep(0, length(r)),
  comparison.type = "representative"
)
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{typeA}{the "from" type that we are interested in, -1 is wildcard}

\item{typeB}{the "to" type that we are interested i, -1 is wildcard}

\item{r}{the series of spatial distances we are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{comparison.type}{what type of points are included in the comparison set.
\itemize{
  \item "representative" if comparison set is representative of the underlying population
  \item "independent" if comparison set is cases/events coming from an indepedent process
}}
}
\value{
tau values for all the distances we looked at
}
\description{
Version of th e \code{get.tau} function that is optimized for
statically typed data. That is data where we want the relationship between
points of type A and points of type B
}
\seealso{
Other get.tau: 
\code{\link{get.tau.bootstrap}()},
\code{\link{get.tau.ci}()},
\code{\link{get.tau.clustsurvey.bootstrap}()},
\code{\link{get.tau.clustsurvey.ci}()},
\code{\link{get.tau.clustsurvey.hh.wts.window}()},
\code{\link{get.tau.clustsurvey.hh.wts}()},
\code{\link{get.tau.clustsurvey.window}()},
\code{\link{get.tau.clustsurvey.wts.bootstrap}()},
\code{\link{get.tau.clustsurvey.wts.ci}()},
\code{\link{get.tau.clustsurvey.wts.window}()},
\code{\link{get.tau.clustsurvey.wts}()},
\code{\link{get.tau.clustsurvey}()},
\code{\link{get.tau.hh.typed.bootstrap}()},
\code{\link{get.tau.hh.typed}()},
\code{\link{get.tau.permute}()},
\code{\link{get.tau.typed.bootstrap.parallel}()},
\code{\link{get.tau.typed.bootstrap}()},
\code{\link{get.tau.typed.clustsurvey.hh.wts}()},
\code{\link{get.tau.typed.clustsurvey.wts}()},
\code{\link{get.tau.typed.permute}()},
\code{\link{get.tau.typed.wts.bootstrap.parallel}()},
\code{\link{get.tau.typed.wts.bootstrap}()},
\code{\link{get.tau.typed.wts.ci}()},
\code{\link{get.tau.typed.wts.permute}()},
\code{\link{get.tau.typed.wts}()},
\code{\link{get.tau.typed}()},
\code{\link{get.tau}()}
}
\author{
Justin Lessler and Henrik Salje
}
\concept{get.tau}
